+++ login-workflow/src/__tests__/screens/Login-test.tsx
@@ -15,18 +15,13 @@ import 'react-native-gesture-handler';
 const Stack = createStackNavigator();
 
 // mock hooks
-jest.mock('src/contexts/AccountUIContext', () => ({
+jest.mock('@pxblue/react-auth-shared', () => ({
+    ...jest.requireActual('@pxblue/react-auth-shared'),
     useAccountUIActions: (): any => ({ dispatch: jest.fn(() => true) }),
     useAccountUIState: jest.fn().mockReturnValue({ login: { transitSuccess: true } }),
     AccountActions: { resetPasswordReset: jest.fn(() => true) },
-}));
-
-jest.mock('src/contexts/AuthUIContextProvider', () => ({
     useAuthUIActions: (): any => ({ dispatch: jest.fn(() => true) }),
     useInjectedUIContext: jest.fn().mockReturnValue({ showSelfRegistration: true }),
-}));
-
-jest.mock('src/contexts/SecurityContextProvider', () => ({
     useSecurityState: jest.fn().mockReturnValue({ rememberMeDetails: { email: undefined, rememberMe: false } }),
 }));