+++ login-workflow/src/components/ErrorState.tsx
@@ -6,7 +6,7 @@
 import React from 'react';
 
 // Hooks
-import { useLanguageLocale } from '../hooks/language-locale-hooks';
+import { useLanguageLocale } from '@pxblue/react-auth-shared';
 
 // Components
 import { EmptyState, wrapIcon } from '@pxblue/react-native-components';
@@ -79,7 +79,7 @@ export const ErrorState: React.FC<ErrorStateProps> = (props) => {
                     IconClass={icon ?? ReportIcon}
                     iconColor={theme.colors.error}
                     title={title ?? t('MESSAGES.FAILURE')}
-                    description={bodyText ?? t('MESSAGES.REQUEST_ERROR')}
+                    description={bodyText !== null ? t(bodyText) : t('MESSAGES.REQUEST_ERROR')}
                 />
             </View>