+++ login-workflow/src/components/PasswordRequirements.tsx
@@ -6,19 +6,24 @@
 import React from 'react';
 
 import { View, StyleProp, ViewStyle } from 'react-native';
-import { useLanguageLocale } from '../hooks/language-locale-hooks';
+
+// Components
+import { RequirementCheck } from '../components/RequirementCheck';
+
+// Shared Auth Logic
 import {
+    // Constants
     SPECIAL_CHAR_REGEX,
     LENGTH_REGEX,
     NUMBERS_REGEX,
     UPPER_CASE_REGEX,
     LOWER_CASE_REGEX,
-} from '../constants/index';
-
-// Components
-import { RequirementCheck } from '../components/RequirementCheck';
-import { useInjectedUIContext } from '../contexts/AuthUIContextProvider';
-import { PasswordRequirement } from '../types/ResetPasswordParams';
+    // Types
+    PasswordRequirement,
+    // Hooks
+    useInjectedUIContext,
+    useLanguageLocale,
+} from '@pxblue/react-auth-shared';
 
 /**
  * @param passwordText  The password to be used for checking.