+++ login-workflow/src/screens/Login.tsx
@@ -5,9 +5,6 @@
 
 import React from 'react';
 
-// Constants
-import { EMAIL_REGEX } from '../constants/index';
-
 // Components
 import { Platform, View, StyleSheet, SafeAreaView, StatusBar, TextInput as ReactTextInput } from 'react-native';
 import { Button, Theme, useTheme } from 'react-native-paper';
@@ -27,11 +24,19 @@ import * as Colors from '@pxblue/colors';
 import { Body1, H6 } from '@pxblue/react-native-components';
 
 // Hooks
-import { useLanguageLocale } from '../hooks/language-locale-hooks';
 import { useNavigation } from '@react-navigation/native';
-import { useAccountUIActions, useAccountUIState } from '../contexts/AccountUIContext';
-import { useInjectedUIContext } from '../contexts/AuthUIContextProvider';
-import { useSecurityState } from '../contexts/SecurityContextProvider';
+
+// Shared Auth Logic
+import {
+    // Constants
+    EMAIL_REGEX,
+    // Hooks
+    useLanguageLocale,
+    useAccountUIActions,
+    useAccountUIState,
+    useInjectedUIContext,
+    useSecurityState,
+} from '@pxblue/react-auth-shared';
 
 /**
  * @ignore
@@ -232,7 +237,7 @@ export const Login: React.FC<LoginProps> = (props) => {
                     uppercase={false}
                     onPress={(): void =>
                         navigation.navigate('PasswordResetCompletion', {
-                            verifyCode: 'DEBUG_VALIDATION_CODE_DEADBEEF',
+                            code: 'DEBUG_VALIDATION_CODE_DEADBEEF',
                         })
                     }
                 >
@@ -254,7 +259,7 @@ export const Login: React.FC<LoginProps> = (props) => {
                     uppercase={false}
                     onPress={(): void =>
                         navigation.navigate('RegistrationInvite', {
-                            validationCode: 'DEBUG_VALIDATION_CODE_DEADBEEF',
+                            code: 'DEBUG_VALIDATION_CODE_DEADBEEF',
                         })
                     }
                 >
@@ -287,6 +292,7 @@ export const Login: React.FC<LoginProps> = (props) => {
                 <SafeAreaView style={[containerStyles.mainContainer, containerStyles.spaceBetween]}>
                     <View style={containerStyles.inputAreas}>
                         <TextInput
+                            testID={'email-text-field'}
                             label={t('LABELS.EMAIL')}
                             value={emailInput}
                             keyboardType={'email-address'}
@@ -300,6 +306,7 @@ export const Login: React.FC<LoginProps> = (props) => {
                             errorText={t('LOGIN.INCORRECT_CREDENTIALS')}
                         />
                         <TextInputSecure
+                            testID={'password-text-field'}
                             ref={confirmPasswordRef}
                             label={t('LABELS.PASSWORD')}
                             value={passwordInput}